pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "d2r-agent"
version = "0.1.0"
description = "Fact-first, retrievable, traceable D2R Q&A agent (MVP)."
requires-python = ">=3.10"

dependencies = [
  "pydantic>=2.6",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "lxml>=5.1",
  "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
